refactor BuyGet limit to be an int to protect it from invalid fractional numbers

+++ src/main/java/com/zimmeren/checkout/BuyGet.java
@@ -5,22 +5,22 @@ public class BuyGet extends Special{
     public float buy;
     public float get;
     public float off;
-    public float limit;
+    public int limit;
 
-    private BuyGet(float buy, float get, float off, float limit) {
+    private BuyGet(float buy, float get, float off, int limit) {
         this.buy = buy;
         this.get = get;
         this.off = off;
         this.limit = limit;
     }
 
-    public static BuyGet of(float buy, float get, float off, float limit) {
+    public static BuyGet of(float buy, float get, float off, int limit) {
         return new BuyGet(buy, get, off, limit);
     }
 
     public float priceOf(float quantity, float stickerPrice) {
         float price = 0f;
-        float rounds = 0f;
+        int rounds = 0;
         while (quantity > 0) {
             if (quantity - buy > 0) {
                 price += buy * stickerPrice;
